pyproject.toml
[project]
name = "macro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mss",
    "pillow",
    "pyqt6",
    "numpy",
]
